+++ src/lib/syntax/color/lch.ts
@@ -106,8 +106,6 @@ export function color2lchToken(token: ColorToken): ColorToken | null {
 }
 
 function lchToken(values: number[]): ColorToken | null {
-    values[2] = values[2];
-
     const chi: Token[] = <Token[]>[
         { typ: EnumToken.NumberTokenType, val: values[0] },
         { typ: EnumToken.NumberTokenType, val: values[1] },